+++ bob-common/src/bob_client.rs
@@ -32,6 +32,7 @@ pub mod b_client {
         local_node_name: NodeName,
 
         operation_timeout: Duration,
+        check_timeout: Duration,
         auth_header: String,
         metrics: BobClientMetrics,
     }
@@ -43,6 +44,7 @@ pub mod b_client {
         pub async fn create(
             node: &Node,
             operation_timeout: Duration,
+            check_timeout: Duration,
             metrics: BobClientMetrics,
             local_node_name: NodeName,
             tls_config: Option<&FactoryTlsConfig>,
@@ -67,11 +69,12 @@ pub mod b_client {
             Ok(Self {
                 client,
                 target_node_name: node.name().clone(),
-                target_node_address: node.address().to_owned(), 
-                local_node_name: local_node_name, 
-                operation_timeout: operation_timeout, 
-                auth_header: auth_header, 
-                metrics: metrics
+                target_node_address: node.address().to_owned(),
+                local_node_name,
+                operation_timeout,
+                check_timeout,
+                auth_header,
+                metrics
             })
         }
 
@@ -155,7 +158,7 @@ pub mod b_client {
             let mut req = Request::new(Null {});
             self.set_credentials(&mut req);
             self.set_node_name(&mut req);
-            self.set_timeout(&mut req);
+            self.set_check_timeout(&mut req);
 
             let node_name = self.target_node_name.to_owned();
             let mut client = self.client.clone();
@@ -243,11 +246,16 @@ pub mod b_client {
         fn set_timeout<T>(&self, r: &mut Request<T>) {
             r.set_timeout(self.operation_timeout);
         }
+
+        fn set_check_timeout<T>(&self, r: &mut Request<T>) {
+            r.set_timeout(self.check_timeout);
+        }
     }
 
     mock! {
         pub BobClient {
-            pub async fn create<'a>(node: &Node, operation_timeout: Duration, metrics: BobClientMetrics, local_node_name: NodeName, tls_config: Option<&'a FactoryTlsConfig>) -> Result<Self, String>;
+            pub async fn create<'a>(node: &Node, operation_timeout: Duration, check_timeout: Duration,
+                                    metrics: BobClientMetrics, local_node_name: NodeName, tls_config: Option<&'a FactoryTlsConfig>) -> Result<Self, String>;
             pub async fn put(&self, key: BobKey, d: BobData, options: PutOptions) -> PutResult;
             pub async fn get(&self, key: BobKey, options: GetOptions) -> GetResult;
             pub async fn ping(&self) -> PingResult;
@@ -315,6 +323,7 @@ pub struct FactoryTlsConfig {
 #[derive(Clone)]
 pub struct Factory {
     operation_timeout: Duration,
+    check_timeout: Duration,
     metrics: Arc<dyn MetricsContainerBuilder + Send + Sync>,
     local_node_name: NodeName,
     tls_config: Option<FactoryTlsConfig>,
@@ -325,12 +334,14 @@ impl Factory {
     #[must_use]
     pub fn new(
         operation_timeout: Duration,
+        check_timeout: Duration,
         metrics: Arc<dyn MetricsContainerBuilder + Send + Sync>,
         local_node_name: NodeName,
         tls_config: Option<FactoryTlsConfig>,
     ) -> Self {
         Factory {
             operation_timeout,
+            check_timeout,
             metrics,
             local_node_name,
             tls_config,
@@ -341,6 +352,7 @@ impl Factory {
         BobClient::create(
             node,
             self.operation_timeout,
+            self.check_timeout,
             metrics,
             self.local_node_name.clone(),
             self.tls_config.as_ref(),